Add/nt/installer/loader: truncation warning. jiraref:VBP-1253
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@106684 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/Installer/Loader/VBoxWindowsAdditions.cpp
@@ -979,7 +979,7 @@ int main()
     * Strip the extension off the module name and construct the arch specific
     * one of the real installer program.
     */
-    DWORD off = cwcExePath - 1;
+    size_t off = cwcExePath - 1;
     while (   off > 0
            && (   wszExePath[off] != '/'
                && wszExePath[off] != '\\'
